Rename reflection calls before remapping
Fixes #906
