+++ .github/workflows/build.yml
@@ -15,7 +15,6 @@ jobs:
       - name: Setup Gradle
         uses: gradle/gradle-build-action@v3
         with:
-          dependency-graph: generate-and-submit
           gradle-home-cache-cleanup: true
           gradle-home-cache-excludes: |
             gradle.properties
